pyproject.toml
[project]
name = "sclaw"
version = "0.0.1"
description = "Easy install, 24/7. Ultra-lightweight secure AI assistant inspired by OpenClaw."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "SClaw Team"}]
keywords = ["ai", "agent", "assistant", "automation", "webui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "aiohttp>=3.9",
    "click>=8.0",
    "pydantic>=2.0",
    "html2text>=2024.2",
    "croniter>=1.3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
semantic = ["chromadb>=0.4"]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
sclaw = "sclaw.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["sclaw*"]
